Deprecate redactedHeaderFields (#32)
Deprecate redactedHeaderFields

### Motivation

For details, see https://github.com/apple/swift-openapi-generator/issues/113.

### Modifications

 This PR deprecates the feature in the current 0.1.x version, so that it can be removed in 0.2.0.

### Result

When adopters use this feature in 0.1.x, they will get a deprecation warning.

### Test Plan

Moved the tests to the deprecated section as well, to be removed before we tag 0.2.0.


Reviewed by: FranzBusch, simonjbeaumont

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (api breakage) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (integration test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-runtime/pull/32
